requirement.txt
git+git://github.com/ildoonet/pytorch-randaugment.git@master
absl-py==0.9.0
asgiref==3.2.10
asn1crypto==1.3.0
astor==0.8.0
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.0
bleach==3.1.4
blinker==1.4
cachetools==3.1.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
conda==4.8.3
conda-build==3.18.11
conda-package-handling==1.6.0
cryptography==2.8
cycler==0.10.0
Cython==0.29.15
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.2
einops==0.3.0
entrypoints==0.3
filelock==3.0.12
future==0.18.2
gast==0.2.2
geffnet==1.0.0
glob2==0.7
google-auth==1.13.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.8.0
idna==2.8
image==1.5.33
imageio==2.8.0
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.1
joblib==0.14.1
json5==0.9.4
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyter-tensorboard==0.2.0
jupyterlab==1.2.6
jupyterlab-server==1.1.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
libarchive-c==2.9
lief==0.9.0
linecache2==1.0.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.3
MiniMock==1.2.8
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
nose==1.3.7
notebook==6.0.3
numpy==1.18.1
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.1.0
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pkginfo==1.5.0.1
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.11.4
psutil==5.7.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycosat==0.6.3
pycparser==2.19
Pygments==2.6.1
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
python-jsonrpc-server==0.4.0
python-language-server==0.36.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==18.1.1
qtconsole==4.7.3
QtPy==1.9.0
RandAugment==0.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
ruamel-yaml==0.15.87
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
six==1.14.0
soupsieve==2.0
sqlparse==0.4.1
tensorboard==2.1.0
tensorboardX==2.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
timm==0.3.1
torch==1.7.0
torchvision==0.6.0a0+82fd1c8
tornado==6.0.4
tqdm==4.42.1
traceback2==1.4.0
traitlets==4.3.3
typing-extensions==3.7.4.3
ujson==4.0.1
unittest2==1.1.0
urllib3==1.25.8
vit-pytorch==0.2.6
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==2.2.0
